Update sql seed file with all relevant tables and convert all Ids to 'long'/'bigint'

+++ src/main/java/com/throughapinhole/photoportfolio/entities/Category.java
@@ -9,7 +9,7 @@ import lombok.Data;
  * Represents a Category of the CMS
  */
 public class Category {
-    private int id;
+    private long id;
     /** Number of items using this Category */
     private int count;
     /** Text description of the Category */
@@ -19,7 +19,7 @@ public class Category {
     /** Text name of the Category */
     private String taxonomy;
     /** ID of the parent category. Has value -1 if there is no parent */
-    private int parent = Utils.CATEGORY_NO_PARENT;
+    private long parent = Utils.CATEGORY_NO_PARENT;
 
     /**
      * Creates a Category with a given description, taxonomy and name
